Fixed memory overflow issue with getPixelColor

+++ Adafruit_NeoPixel.cpp
@@ -1634,9 +1634,9 @@ uint32_t Adafruit_NeoPixel::getPixelColor(uint16_t n) const {
       // value used when setting the pixel color, but there will always be
       // some error -- those bits are simply gone.  Issue is most
       // pronounced at low brightness levels.
-      return (((uint32_t)(p[rOffset] << 8) / brightness) << 16) |
-             (((uint32_t)(p[gOffset] << 8) / brightness) <<  8) |
-             ( (uint32_t)(p[bOffset] << 8) / brightness       );
+      return ((uint32_t)((uint16_t)(p[rOffset] << 8) / brightness) << 16) |
+             ((uint32_t)((uint16_t)(p[gOffset] << 8) / brightness) <<  8) |
+              (uint32_t)((uint16_t)(p[bOffset] << 8) / brightness       );
     } else {
       // No brightness adjustment has been made -- return 'raw' color
       return ((uint32_t)p[rOffset] << 16) |
@@ -1646,10 +1646,10 @@ uint32_t Adafruit_NeoPixel::getPixelColor(uint16_t n) const {
   } else {                 // Is RGBW-type device
     p = &pixels[n * 4];
     if(brightness) { // Return scaled color
-      return (((uint32_t)(p[wOffset] << 8) / brightness) << 24) |
-             (((uint32_t)(p[rOffset] << 8) / brightness) << 16) |
-             (((uint32_t)(p[gOffset] << 8) / brightness) <<  8) |
-             ( (uint32_t)(p[bOffset] << 8) / brightness       );
+      return ((uint32_t)((uint16_t)(p[wOffset] << 8) / brightness) << 24) |
+             ((uint32_t)((uint16_t)(p[rOffset] << 8) / brightness) << 16) |
+             ((uint32_t)((uint16_t)(p[gOffset] << 8) / brightness) <<  8) |
+              (uint32_t)((uint16_t)(p[bOffset] << 8) / brightness       );
     } else { // Return raw color
       return ((uint32_t)p[wOffset] << 24) |
              ((uint32_t)p[rOffset] << 16) |
